requirements.txt
streamlit==1.29.0
openai==2.8.1
python-dotenv==1.0.0
requests==2.31.0
google-search-results==2.4.2
pdfplumber==0.10.3
python-docx==1.1.0
fpdf2==2.7.9
Pillow